Replace original CongressMembers class with ProPublicaCongressAdapter class for accessing API but still having the ability to save congress member info to PostgreSQL. Create migration and for new version of CongressMembers model and alter seed file to send current members into the local db for search in association with their name such that API calls will be made for individual members using their politician ID.
