+++ app/controllers/users_controller.rb
@@ -39,9 +39,9 @@ class UsersController < ApplicationController
       and status = 'unanswered'")
 
     # members are not currently sorted by alpha order beyond their last initial
-    @house_members = CongressMember.new(endpoint: "member_list", branch: "house")
+    @house_members = ProPublicaCongressAdapter.new(endpoint: "member_list", branch: "house")
     @house_members = @house_members.members_basic_details
-    @senate_members = CongressMember.new(endpoint: "member_list", branch: "senate")
+    @senate_members = ProPublicaCongressAdapter.new(endpoint: "member_list", branch: "senate")
     @senate_members = @senate_members.members_basic_details
 
     render 'users/hub'
