+++ db/seeds.rb
@@ -9,8 +9,9 @@
 require 'faker'
 require 'uri'
 require 'net/http'
+require 'openssl'
 
-# SEEDING USERS
+# USERS
 max_users = 15
 users_needed = max_users - User.count
 
@@ -39,7 +40,7 @@ end
 
 User.create(username: "MikeT", email: "m@m.com", password: "pass", avatar: "https://github.com/MikeTarkington/mike_tarkington_site/blob/master/square_profile_img_hr.png?raw=true")
 
-# SEEDING BASE QUIZ
+# BASE QUIZ QUESTIONS
 max_quizzes = 15
 quizzes_needed = max_quizzes - Quiz.count
 
@@ -55,7 +56,7 @@ quizzes_needed.times do
 
 end
 
-# SEEDING QUIZ RESPONSES
+# QUIZ RESPONSES
 max_responses = 15
 responses_needed = max_responses - Response.count
 
@@ -74,7 +75,7 @@ responses_needed.times do
 
 end
 
-# SEEDING FRIENDSHIPS
+# FRIENDSHIPS
 max_friendships = 100
 friendships_needed = max_friendships - Friendship.count
 
@@ -101,7 +102,7 @@ end
   Friendship.create(user_id: rand(1..15), friended_user: 16, status: ["approved", "unanswered"].sample)
 end
 
-# SEEDING ENTRIES
+# ENTRIES
 max_entries = 50
 entries_needed = max_entries - Entry.count
 
@@ -120,7 +121,7 @@ end
   Entry.create(user_id: 16, title: Faker::GameOfThrones.character, body: Faker::ChuckNorris.fact, entry_type: ["profile_post", "memo", "group_page"].sample)
 end
 
-# SEEDING COMMENTS
+# COMMENTS
 max_comments = 50
 comments_needed = max_comments - Comment.count
 
@@ -146,3 +147,20 @@ mikes_posts.each do |post|
     Comment.create(comment_data)
   end
 end
+
+# CONGRESS MEMBERS
+def congress_branch_seed(branch)
+  instantiate_congress = ProPublicaCongressAdapter.new(endpoint: "member_list", branch: branch)
+  current_members = instantiate_congress.parsed_response["results"][0]["members"]
+  current_member = {}
+  current_members.each_with_index do |member, i|
+    current_member[:f_name] = current_members[i]["first_name"]
+    current_member[:l_name] = current_members[i]["last_name"]
+    current_member[:politician_id] = current_members[i]["id"]
+
+    CongressMember.create(current_member)
+  end
+end
+
+congress_branch_seed("house")
+congress_branch_seed("senate")
