+++ test/jobs/update_congress_job_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class UpdateCongressJobTest < ActiveJob::TestCase
+  # test "the truth" do
+  #   assert true
+  # end
+end
